Create a spatial attention block
