+++ monai/networks/blocks/transformerblock.py
@@ -13,8 +13,8 @@ from __future__ import annotations
 
 import torch.nn as nn
 
+from monai.networks.blocks.attention import SABlock
 from monai.networks.blocks.mlp import MLPBlock
-from monai.networks.blocks.selfattention import SABlock
 
 
 class TransformerBlock(nn.Module):
