+++ monai/networks/nets/autoencoderkl.py
@@ -19,7 +19,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from monai.networks.blocks import Convolution, Upsample
-from monai.networks.blocks.selfattention import SABlock
+from monai.networks.blocks.attention import SpatialAttentionBlock
 from monai.utils import ensure_tuple_rep, optional_import
 
 Rearrange, _ = optional_import("einops.layers.torch", name="Rearrange")
@@ -144,61 +144,6 @@ class AEKLResBlock(nn.Module):
         return x + h
 
 
-class AttentionBlock(nn.Module):
-    """Perform spatial self-attention on the input tensor.
-
-    The input tensor is reshaped to B x (x_dim * y_dim [ * z_dim]) x C, where C is the number of channels.
-
-    Args:
-        spatial_dims: number of spatial dimensions, could be 1, 2, or 3.
-        num_channels: number of input channels. Must be divisible by num_head_channels.
-        num_head_channels: number of channels per head.
-    """
-
-    def __init__(
-        self,
-        spatial_dims: int,
-        num_channels: int,
-        num_head_channels: int | None = None,
-        norm_num_groups: int = 32,
-        norm_eps: float = 1e-6,
-    ) -> None:
-        super().__init__()
-
-        self.spatial_dims = spatial_dims
-        self.norm = nn.GroupNorm(num_groups=norm_num_groups, num_channels=num_channels, eps=norm_eps, affine=True)
-        # check num_head_channels is divisible by num_channels
-        if num_head_channels is not None and num_channels % num_head_channels != 0:
-            raise ValueError("num_channels must be divisible by num_head_channels")
-        num_heads = num_channels // num_head_channels if num_head_channels is not None else 1
-
-        self.attn = SABlock(hidden_size=num_channels, num_heads=num_heads, qkv_bias=True)
-
-    def forward(self, x: torch.Tensor):
-        residual = x
-
-        if self.spatial_dims == 1:
-            h = x.shape[2]
-            rearrange_input = Rearrange("b c h -> b h c")
-            rearrange_output = Rearrange("b h c -> b c h", h=h)
-        if self.spatial_dims == 2:
-            h, w = x.shape[2], x.shape[3]
-            rearrange_input = Rearrange("b c h w -> b (h w) c")
-            rearrange_output = Rearrange("b (h w) c -> b c h w", h=h, w=w)
-        if self.spatial_dims == 3:
-            h, w, d = x.shape[2], x.shape[3], x.shape[4]
-            rearrange_input = Rearrange("b c h w d -> b (h w d) c")
-            rearrange_output = Rearrange("b (h w d) c -> b c h w d", h=h, w=w, d=d)
-
-        x = self.norm(x)
-        x = rearrange_input(x)  # B x (x_dim * y_dim [ * z_dim]) x C
-
-        x = self.attn(x)
-        x = rearrange_output(x)  # B x  x C x x_dim * y_dim * [z_dim]
-        x = x + residual
-        return x
-
-
 class Encoder(nn.Module):
     """
     Convolutional cascade that downsamples the image into a spatial latent space.
@@ -271,7 +216,7 @@ class Encoder(nn.Module):
                 input_channel = output_channel
                 if attention_levels[i]:
                     blocks.append(
-                        AttentionBlock(
+                        SpatialAttentionBlock(
                             spatial_dims=spatial_dims,
                             num_channels=input_channel,
                             norm_num_groups=norm_num_groups,
@@ -294,7 +239,7 @@ class Encoder(nn.Module):
             )
 
             blocks.append(
-                AttentionBlock(
+                SpatialAttentionBlock(
                     spatial_dims=spatial_dims,
                     num_channels=channels[-1],
                     norm_num_groups=norm_num_groups,
@@ -401,7 +346,7 @@ class Decoder(nn.Module):
                 )
             )
             blocks.append(
-                AttentionBlock(
+                SpatialAttentionBlock(
                     spatial_dims=spatial_dims,
                     num_channels=reversed_block_out_channels[0],
                     norm_num_groups=norm_num_groups,
@@ -440,7 +385,7 @@ class Decoder(nn.Module):
 
                 if reversed_attention_levels[i]:
                     blocks.append(
-                        AttentionBlock(
+                        SpatialAttentionBlock(
                             spatial_dims=spatial_dims,
                             num_channels=block_in_ch,
                             norm_num_groups=norm_num_groups,
